test(#664): C6.1-index — drain ALLOWED_FULL_INDEX to 0, relocate full-index e2e tests to integration
Relocate the 13 remaining ALLOWED_FULL_INDEX offenders to tests/integration/.
Each calls akmIndex({ full: true }) inside a test() body because the on-disk
FTS/vector rebuild IS the subject (ensureIndex staleness, --clean, indexer
rejection, semantic-search/secret/session indexing e2e, source scoping, improve
over a real indexed stash) — genuine process/FTS-bound integration tests, not
planner-only candidates for seedEntries. Moving them out of unit scope is the
design-sanctioned path (§3 residual integration cases).

- git mv all 13 into tests/integration/ (preserving commands/ and workflows/
  subdir structure); fix relative import depth (+1 ../).
- Empty ALLOWED_FULL_INDEX and lower UNIT_PURITY_BASELINE 13 -> 0. The unit-tier
  purity ratchet is now fully drained: ALLOWED_SERVE, ALLOWED_SPAWN, and
  ALLOWED_FULL_INDEX are all empty and the baseline is 0.
- Update the 8 relocated paths referenced in lint-tests-isolation.ts's
  ALLOWED_FILES to their new tests/integration/ locations (set membership and
  ALLOWLIST_RATCHET_BASELINE unchanged — a rename, not a removal).

Verified: bun run lint (incl. both ratchet lints), bunx tsc --noEmit,
all 13 relocated suites green at their new location, and the
unit-purity + isolation ratchet meta-tests pass.
