+++ tests/integration/commands/improve-memory.test.ts
@@ -2,18 +2,18 @@ import { afterEach, beforeEach, describe, expect, spyOn, test } from "bun:test";
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
-import type { AkmDistillResult } from "../../src/commands/improve/distill";
-import { akmImprove } from "../../src/commands/improve/improve";
-import type { AkmReflectResult } from "../../src/commands/improve/reflect";
-import type { Proposal } from "../../src/commands/proposal/validators/proposals";
-import { akmSearch } from "../../src/commands/read/search";
-import { saveConfig } from "../../src/core/config/config";
-import { appendEvent, readEvents } from "../../src/core/events";
-import { setQuiet } from "../../src/core/warn";
-import type { GraphExtractionResult } from "../../src/indexer/graph/graph-extraction";
-import { akmIndex } from "../../src/indexer/indexer";
-import type { MemoryInferenceResult } from "../../src/indexer/passes/memory-inference";
-import { getWebsiteCachePaths } from "../../src/sources/website-ingest";
+import type { AkmDistillResult } from "../../../src/commands/improve/distill";
+import { akmImprove } from "../../../src/commands/improve/improve";
+import type { AkmReflectResult } from "../../../src/commands/improve/reflect";
+import type { Proposal } from "../../../src/commands/proposal/validators/proposals";
+import { akmSearch } from "../../../src/commands/read/search";
+import { saveConfig } from "../../../src/core/config/config";
+import { appendEvent, readEvents } from "../../../src/core/events";
+import { setQuiet } from "../../../src/core/warn";
+import type { GraphExtractionResult } from "../../../src/indexer/graph/graph-extraction";
+import { akmIndex } from "../../../src/indexer/indexer";
+import type { MemoryInferenceResult } from "../../../src/indexer/passes/memory-inference";
+import { getWebsiteCachePaths } from "../../../src/sources/website-ingest";
 
 const tempDirs: string[] = [];
