+++ tests/integration/commands/improve/improve-dry-run-side-effects.test.ts
@@ -25,9 +25,9 @@ import { createHash } from "node:crypto";
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
-import { akmImprove } from "../../../src/commands/improve/improve";
-import { saveConfig } from "../../../src/core/config/config";
-import { akmIndex } from "../../../src/indexer/indexer";
+import { akmImprove } from "../../../../src/commands/improve/improve";
+import { saveConfig } from "../../../../src/core/config/config";
+import { akmIndex } from "../../../../src/indexer/indexer";
 
 const TIMEOUT_MS = 15_000;
