+++ tests/integration/commands/improve/improve-sync.test.ts
@@ -21,12 +21,12 @@
 import { afterEach, beforeEach, describe, expect, mock, test } from "bun:test";
 import fs from "node:fs";
 import path from "node:path";
-import { akmImprove } from "../../../src/commands/improve/improve";
-import type { AkmConfig } from "../../../src/core/config/config";
-import { saveConfig } from "../../../src/core/config/config";
-import { akmIndex } from "../../../src/indexer/indexer";
-import type { SaveGitStashResult } from "../../../src/sources/providers/git";
-import { type Cleanup, withIsolatedAkmStorage } from "../../_helpers/sandbox";
+import { akmImprove } from "../../../../src/commands/improve/improve";
+import type { AkmConfig } from "../../../../src/core/config/config";
+import { saveConfig } from "../../../../src/core/config/config";
+import { akmIndex } from "../../../../src/indexer/indexer";
+import type { SaveGitStashResult } from "../../../../src/sources/providers/git";
+import { type Cleanup, withIsolatedAkmStorage } from "../../../_helpers/sandbox";
 
 const TIMEOUT_MS = 20_000;
