+++ tests/integration/issue-36-repro.test.ts
@@ -12,12 +12,12 @@ import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
-import { akmSearch } from "../src/commands/read/search";
-import { saveConfig } from "../src/core/config/config";
-import { closeDatabase, getAllEntries, openDatabase, searchFts } from "../src/indexer/db/db";
-import { akmIndex } from "../src/indexer/indexer";
-import type { SourceSearchHit } from "../src/sources/types";
-import { type Cleanup, sandboxStashDir, sandboxXdgCacheHome, sandboxXdgConfigHome } from "./_helpers/sandbox";
+import { akmSearch } from "../../src/commands/read/search";
+import { saveConfig } from "../../src/core/config/config";
+import { closeDatabase, getAllEntries, openDatabase, searchFts } from "../../src/indexer/db/db";
+import { akmIndex } from "../../src/indexer/indexer";
+import type { SourceSearchHit } from "../../src/sources/types";
+import { type Cleanup, sandboxStashDir, sandboxXdgCacheHome, sandboxXdgConfigHome } from "../_helpers/sandbox";
 
 // ── Helpers ─────────────────────────────────────────────────────────────────
