+++ tests/integration/source.test.ts
@@ -2,13 +2,13 @@ import { afterAll, afterEach, beforeEach, describe, expect, test } from "bun:tes
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
-import { akmSearch } from "../src/commands/read/search";
-import { akmShowUnified as akmShow } from "../src/commands/read/show";
-import { akmInit } from "../src/commands/sources/init";
-import { resetConfigCache, saveConfig } from "../src/core/config/config";
-import { getBinDir, getConfigPath } from "../src/core/paths";
-import { akmIndex } from "../src/indexer/indexer";
-import type { SearchHit, SourceSearchHit } from "../src/sources/types";
+import { akmSearch } from "../../src/commands/read/search";
+import { akmShowUnified as akmShow } from "../../src/commands/read/show";
+import { akmInit } from "../../src/commands/sources/init";
+import { resetConfigCache, saveConfig } from "../../src/core/config/config";
+import { getBinDir, getConfigPath } from "../../src/core/paths";
+import { akmIndex } from "../../src/indexer/indexer";
+import type { SearchHit, SourceSearchHit } from "../../src/sources/types";
 
 const createdTmpDirs: string[] = [];
